+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.0-rc.10 : 02.08.2024
 
+- **Added**: `AliasStrategy` for overriding property name on target object during `setattr()`
+- **Changed**:`field_strategy` now can be also an instance of `BaseStrategy`
 - **Fixed**: Fixed calling population methods when declared in form
 
 ## 1.0.0-rc.9 : 24.03.2023